backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
drf-spectacular==0.28.0
psycopg[binary]==3.2.9
celery==5.5.3
redis==6.2.0
meilisearch==0.37.0
Pillow==11.2.1
django-filter==25.1
django-storages==1.14.6
boto3==1.38.46
python-dotenv==1.1.0
gunicorn==23.0.0
